requirements.txt
asgiref==3.3.1
automium==0.2.6
automium-web==0.1.1
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==8.0.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.1
DateTime==4.3
defusedxml==0.7.0rc1
Django==3.1.4
django-braces==1.14.0
django-cors-headers==3.6.0
django-filter==2.4.0
django-oauth-toolkit==1.3.3
django-rest-framework-social-oauth2==1.1.0
django-templated-mail==1.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
Flask==2.0.1
gunicorn==20.1.0
idna==2.10
itsdangerous==2.0.1
itypes==1.2.0
Jinja2==3.0.1
legacy==0.1.6
MarkupSafe==2.0.1
mysql-connector-python==8.0.22
mysqlclient==2.0.2
oauthlib==3.1.0
protobuf==3.14.0
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.2
Werkzeug==2.0.1
whitenoise
zope.interface==5.4.0